[3AP] - Add new model, move old ones
