+++ src/main/java/ch/aaap/ca/be/medicalsupplies/model/MSGenericNameRow.java
@@ -1,4 +1,4 @@
-package ch.aaap.ca.be.medicalsupplies.data;
+package ch.aaap.ca.be.medicalsupplies.model;
 
 import org.apache.commons.lang3.builder.ToStringBuilder;
 
@@ -12,7 +12,7 @@ public class MSGenericNameRow {
     private final String category4;
 
     public MSGenericNameRow(Long id, String name, String category1, String category2, String category3,
-            String category4) {
+                            String category4) {
         this.id = id;
         this.name = name;
         this.category1 = category1;
@@ -49,4 +49,23 @@ public class MSGenericNameRow {
     public String toString() {
         return ToStringBuilder.reflectionToString(this);
     }
+
+
+    //This is probably the easiest solution but we need duplicates for other steps
+//    @Override
+//    public boolean equals(Object obj) {
+//        if (obj == null) {
+//            return false;
+//        }
+//        if (!(obj instanceof MSGenericNameRow)) {
+//            return false;
+//        }
+//
+//        return this.getName().equals(((MSGenericNameRow) obj).getName());
+//    }
+//
+//    @Override
+//    public int hashCode() {
+//        return getName().hashCode();
+//    }
 }
